+++ app/maingui.py
@@ -1,20 +1,17 @@
 import os
 import shutil
-import sys
 import time
-import traceback
 from datetime import datetime
 from threading import Event
 
 import clr
-from PyQt6.QtCore import (QObject, QRunnable, Qt, QThreadPool, pyqtSignal,
-                        pyqtSlot)
+from PyQt6.QtCore import (Qt, QThreadPool)
 from PyQt6.QtGui import QIcon
 from PyQt6.QtWidgets import QFileDialog, QMainWindow
 
 from downloader import Downloader
 from get_url import url_window
-from misc import DilalogBox, Miscellaneous
+from misc import DilalogBox, Miscellaneous,Worker
 from url_gen import url_generator
 from utls import open_browser
 
@@ -22,79 +19,7 @@ script_dir = os.path.dirname(os.path.abspath(__file__))
 dll_path = os.path.join(script_dir,r"data\System.Management.Automation.dll")
 clr.AddReference(dll_path)
 
-class WorkerSignals(QObject):
-    '''
-    Defines the signals available from a running worker thread.
-
-    Supported signals are:
-
-    finished
-        No data
-
-    error
-        tuple (exctype, value, traceback.format_exc() )
-
-    result
-        object data returned from processing, anything
-
-    progress
-        int indicating % progress
-
-    '''
-    started = pyqtSignal()
-    finished = pyqtSignal()
-    error = pyqtSignal(tuple)
-    result = pyqtSignal(object)
-    cur_progress = pyqtSignal(int)
-    main_progress = pyqtSignal(int)
-
-
-class Worker(QRunnable):
-    '''
-    Worker thread
-
-    Inherits from QRunnable to handler worker thread setup, signals and wrap-up.
-
-    :param callback: The function callback to run on this worker thread. Supplied args and
-                    kwargs will be passed through to the runner.
-    :type callback: function
-    :param args: Arguments to pass to the callback function
-    :param kwargs: Keywords to pass to the callback function
-
-    '''
-
-    def __init__(self, fn, *args, **kwargs):
-        super(Worker, self).__init__()
-
-        # Store constructor arguments (re-used for processing)
-        self.fn = fn
-        self.args = args
-        self.kwargs = kwargs
-        self.signals = WorkerSignals()
-
-        # Add the callback to our kwargs
-        self.kwargs['progress_current'] = self.signals.cur_progress
-        self.kwargs['progress_main'] = self.signals.main_progress
-
-    @pyqtSlot()
-    def run(self):
-        '''Initialise the runner function with passed args, kwargs.'''
-        # Retrieve args/kwargs here; and fire processing using them
-        try:
-            self.signals.started.emit()
-            result = self.fn(*self.args, **self.kwargs)
-        except:
-            traceback.print_exc()
-            exctype, value = sys.exc_info()[:2]
-            self.signals.error.emit((exctype, value, traceback.format_exc()))
-        else:
-            # Return the result of the processing
-            self.signals.result.emit(result)
-        finally:
-            self.signals.finished.emit()  # Done
-
-
-class MainWindowGui(Miscellaneous):
+class App(Miscellaneous):
     def __init__(self):
         super().__init__()
         self.threadpool = QThreadPool()
@@ -103,144 +28,6 @@ class MainWindowGui(Miscellaneous):
         self.stop = Event()
         self.ignore_ver = False
         self.all_dependencies = False
-
-    def setupUi(self, *args, **kwargs):
-        Miscellaneous.setupUi(self, *args, **kwargs)
-        self.set_bar_0()
-        self.show_bar(False)
-        self.pushButton.clicked.connect(self.openWindow)
-        self.stop_btn.clicked.connect(self.stop_func)
-        self.stop_btn.hide()
-        self.actioninstall_From_File.triggered.connect(
-            self.standalone_installer)
-        self.actionclear_cache.triggered.connect(self.clear_cache)
-        self.actionCheck_For_Updates.triggered.connect(lambda: open_browser(
-            'https://github.com/m-jishnu/alt-app-installer/releases'))
-        self.actionAbout.triggered.connect(lambda: open_browser(
-            'https://github.com/m-jishnu/alt-app-installer'))
-        self.actionHelp.triggered.connect(lambda: open_browser(
-            'https://discord.com/invite/9eeN2Wve4T'))
-        self.actionOpen_Logs.triggered.connect(self.open_Logs)
-        self.actionDownloads.triggered.connect(self.open_downloads)
-        self.actionIgnore_Latest_Version.triggered.connect(self.ignore_version)
-        self.actionIgnore_All_filters.triggered.connect(
-            self.ignore_All_filters)
-        self.actionInstall_using_url.triggered.connect(self.install_url)
-
-    def ignore_version(self):
-        if self.actionIgnore_Latest_Version.isChecked():
-            self.ignore_ver = True
-        else:
-            self.ignore_ver = False
-
-    def ignore_All_filters(self):
-        if self.actionIgnore_All_filters.isChecked():
-            self.all_dependencies = True
-            self.actionIgnore_Latest_Version.setChecked(True)
-            self.actionIgnore_Latest_Version.setEnabled(False)
-        else:
-            self.all_dependencies = False
-            self.actionIgnore_Latest_Version.setChecked(False)
-            self.actionIgnore_Latest_Version.setEnabled(True)
-
-    def open_Logs(self):
-        path = 'log.txt'
-        if os.path.exists(path):
-            os.startfile(path)
-        else:
-            self.show_error_popup(txt="No Logs Found!")
-
-    def clear_cache(self):
-        def remove_file():
-            def remove_(path, mode='file'):
-                if mode == 'file':
-                    if os.path.exists(path):
-                        os.remove(path)
-                    else:
-                        pass
-
-                elif mode == 'dir':
-                    shutil.rmtree(path)
-
-            remove_('log.txt')
-            try:
-                remove_('downloads', 'dir')
-            except FileNotFoundError:
-                print("No Downloads Found!")
-
-        worker = Worker(lambda *ars, **kwargs: remove_file())
-        worker.signals.error.connect(lambda arg: self.error_handler(
-            arg, normal=False, msg="Failed To Clear Cache Files!", critical=False))
-
-        self.threadpool.start(worker)
-        worker.signals.result.connect(lambda: self.show_success_popup(
-            text="Cache Files Cleared Successfully!"))
-
-    def open_downloads(self):
-        path = os.path.realpath("./downloads")
-        if os.path.exists(path):
-            os.startfile(path)
-        else:
-            self.show_error_popup(txt="No Downloads Found!")
-
-    # standalone installer for predownloaded files
-    def standalone_installer(self):
-        def error(arg):
-            self.pushButton.setEnabled(True)
-            self.show_bar(False)
-
-        fname = QFileDialog.getOpenFileNames()[0]
-        if fname:
-            worker = Worker(lambda **kwargs: self.install(fname[0], **kwargs))
-            worker.signals.cur_progress.connect(self.cur_Progress)
-            worker.signals.main_progress.connect(self.main_Progress)
-            worker.signals.result.connect(self.run_success)
-            worker.signals.error.connect(error)
-            self.threadpool.start(worker)
-            self.show_bar(True)
-            self.pushButton.setEnabled(False)
-            self.menuDependencies.setEnabled(False)
-            self.actionclear_cache.setEnabled(False)
-            self.actioninstall_From_File.setEnabled(False)
-            self.actionInstall_using_url.setEnabled(False)
-            # if the app selector window is open closing it
-            try:
-                self.window.close()
-                self.window.deleteLater()
-                del self.window
-            except:
-                pass
-
-    def install_url(self):
-        window = DilalogBox()
-        window.closed.connect(self.parser)
-        window.exec()
-
-    def openWindow(self):
-        # close event for the new window
-        def close(event):
-            self.window.deleteLater()
-            del self.window
-            event.accept()
-
-        try:
-            self.window  # checking if self.window already exist
-        except:
-            self.window = False  # if not set it to false aka the window is not open
-
-        if self.window:  # if it has value then change focus to the already open window
-            self.window.setWindowState(self.window.windowState(
-            ) & ~Qt.WindowState.WindowMinimized | Qt.WindowState.WindowActive)  # if minimized then unminimize
-            self.window.activateWindow()  # set focus to the currently open window
-        else:  # open a new window
-            self.window = QMainWindow()
-            self.window.setWindowIcon(QIcon('./data/images/search.png'))
-            search_app = url_window()
-            search_app.setupUi(self.window)
-            # overding the new window close event for proper cleanup
-            self.window.closeEvent = close
-            self.window.show()
-            search_app.closed.connect(self.parser)
 
     def parser(self, arg):
         self.stop.clear()
@@ -403,3 +190,146 @@ class MainWindowGui(Miscellaneous):
         worker.signals.error.connect(
             lambda arg: self.error_handler(arg, normal=False))
         self.threadpool.start(worker)
+
+
+class MainWindowGui(App):
+    def __init__(self):
+        super().__init__()
+
+    def setupUi(self, *args, **kwargs):
+        Miscellaneous.setupUi(self, *args, **kwargs)
+        self.set_bar_0()
+        self.show_bar(False)
+        self.pushButton.clicked.connect(self.openWindow)
+        self.stop_btn.clicked.connect(self.stop_func)
+        self.stop_btn.hide()
+        self.actioninstall_From_File.triggered.connect(
+            self.standalone_installer)
+        self.actionclear_cache.triggered.connect(self.clear_cache)
+        self.actionCheck_For_Updates.triggered.connect(lambda: open_browser(
+            'https://github.com/m-jishnu/alt-app-installer/releases'))
+        self.actionAbout.triggered.connect(lambda: open_browser(
+            'https://github.com/m-jishnu/alt-app-installer'))
+        self.actionHelp.triggered.connect(lambda: open_browser(
+            'https://discord.com/invite/9eeN2Wve4T'))
+        self.actionOpen_Logs.triggered.connect(self.open_Logs)
+        self.actionDownloads.triggered.connect(self.open_downloads)
+        self.actionIgnore_Latest_Version.triggered.connect(self.ignore_version)
+        self.actionIgnore_All_filters.triggered.connect(
+            self.ignore_All_filters)
+        self.actionInstall_using_url.triggered.connect(self.install_url)
+
+    def ignore_version(self):
+        if self.actionIgnore_Latest_Version.isChecked():
+            self.ignore_ver = True
+        else:
+            self.ignore_ver = False
+
+    def ignore_All_filters(self):
+        if self.actionIgnore_All_filters.isChecked():
+            self.all_dependencies = True
+            self.actionIgnore_Latest_Version.setChecked(True)
+            self.actionIgnore_Latest_Version.setEnabled(False)
+        else:
+            self.all_dependencies = False
+            self.actionIgnore_Latest_Version.setChecked(False)
+            self.actionIgnore_Latest_Version.setEnabled(True)
+
+    def open_Logs(self):
+        path = 'log.txt'
+        if os.path.exists(path):
+            os.startfile(path)
+        else:
+            self.show_error_popup(txt="No Logs Found!")
+
+    def clear_cache(self):
+        def remove_file():
+            def remove_(path, mode='file'):
+                if mode == 'file':
+                    if os.path.exists(path):
+                        os.remove(path)
+                    else:
+                        pass
+
+                elif mode == 'dir':
+                    shutil.rmtree(path)
+
+            remove_('log.txt')
+            try:
+                remove_('downloads', 'dir')
+            except FileNotFoundError:
+                print("No Downloads Found!")
+
+        worker = Worker(lambda *ars, **kwargs: remove_file())
+        worker.signals.error.connect(lambda arg: self.error_handler(
+            arg, normal=False, msg="Failed To Clear Cache Files!", critical=False))
+
+        self.threadpool.start(worker)
+        worker.signals.result.connect(lambda: self.show_success_popup(
+            text="Cache Files Cleared Successfully!"))
+
+    def open_downloads(self):
+        path = os.path.realpath("./downloads")
+        if os.path.exists(path):
+            os.startfile(path)
+        else:
+            self.show_error_popup(txt="No Downloads Found!")
+
+    # standalone installer for predownloaded files
+    def standalone_installer(self):
+        def error(arg):
+            self.pushButton.setEnabled(True)
+            self.show_bar(False)
+
+        fname = QFileDialog.getOpenFileNames()[0]
+        if fname:
+            worker = Worker(lambda **kwargs: self.install(fname[0], **kwargs))
+            worker.signals.cur_progress.connect(self.cur_Progress)
+            worker.signals.main_progress.connect(self.main_Progress)
+            worker.signals.result.connect(self.run_success)
+            worker.signals.error.connect(error)
+            self.threadpool.start(worker)
+            self.show_bar(True)
+            self.pushButton.setEnabled(False)
+            self.menuDependencies.setEnabled(False)
+            self.actionclear_cache.setEnabled(False)
+            self.actioninstall_From_File.setEnabled(False)
+            self.actionInstall_using_url.setEnabled(False)
+            # if the app selector window is open closing it
+            try:
+                self.window.close()
+                self.window.deleteLater()
+                del self.window
+            except:
+                pass
+
+    def install_url(self):
+        window = DilalogBox()
+        window.closed.connect(self.parser)
+        window.exec()
+
+    def openWindow(self):
+        # close event for the new window
+        def close(event):
+            self.window.deleteLater()
+            del self.window
+            event.accept()
+
+        try:
+            self.window  # checking if self.window already exist
+        except:
+            self.window = False  # if not set it to false aka the window is not open
+
+        if self.window:  # if it has value then change focus to the already open window
+            self.window.setWindowState(self.window.windowState(
+            ) & ~Qt.WindowState.WindowMinimized | Qt.WindowState.WindowActive)  # if minimized then unminimize
+            self.window.activateWindow()  # set focus to the currently open window
+        else:  # open a new window
+            self.window = QMainWindow()
+            self.window.setWindowIcon(QIcon('./data/images/search.png'))
+            search_app = url_window()
+            search_app.setupUi(self.window)
+            # overding the new window close event for proper cleanup
+            self.window.closeEvent = close
+            self.window.show()
+            search_app.closed.connect(self.parser)
