+++ app/misc.py
@@ -1,9 +1,12 @@
 import os
+import sys
+import traceback
 from datetime import datetime
 
-from PyQt6.QtGui import QIcon,QPixmap
-from PyQt6.QtWidgets import QMessageBox,QDialog
-from PyQt6 import QtCore,QtWidgets
+from PyQt6 import QtCore, QtWidgets
+from PyQt6.QtCore import QObject, QRunnable, pyqtSignal, pyqtSlot
+from PyQt6.QtGui import QIcon, QPixmap
+from PyQt6.QtWidgets import QDialog, QMessageBox
 
 from gui import Ui_MainProgram
 
@@ -168,6 +171,7 @@ class Miscellaneous(Ui_MainProgram):
             event.accept()
         else:
             event.ignore()
+
 class DilalogBox(QDialog):
 
     closed = QtCore.pyqtSignal(object)
@@ -210,3 +214,73 @@ class DilalogBox(QDialog):
             Form.close()
 
         self.install_link_ok_btn.clicked.connect(current_url)
+
+class WorkerSignals(QObject):
+    '''
+    Defines the signals available from a running worker thread.
+
+    Supported signals are:
+
+    finished
+        No data
+
+    error
+        tuple (exctype, value, traceback.format_exc() )
+
+    result
+        object data returned from processing, anything
+
+    progress
+        int indicating % progress
+
+    '''
+    started = pyqtSignal()
+    finished = pyqtSignal()
+    error = pyqtSignal(tuple)
+    result = pyqtSignal(object)
+    cur_progress = pyqtSignal(int)
+    main_progress = pyqtSignal(int)
+
+class Worker(QRunnable):
+    '''
+    Worker thread
+
+    Inherits from QRunnable to handler worker thread setup, signals and wrap-up.
+
+    :param callback: The function callback to run on this worker thread. Supplied args and
+                    kwargs will be passed through to the runner.
+    :type callback: function
+    :param args: Arguments to pass to the callback function
+    :param kwargs: Keywords to pass to the callback function
+
+    '''
+
+    def __init__(self, fn, *args, **kwargs):
+        super(Worker, self).__init__()
+
+        # Store constructor arguments (re-used for processing)
+        self.fn = fn
+        self.args = args
+        self.kwargs = kwargs
+        self.signals = WorkerSignals()
+
+        # Add the callback to our kwargs
+        self.kwargs['progress_current'] = self.signals.cur_progress
+        self.kwargs['progress_main'] = self.signals.main_progress
+
+    @pyqtSlot()
+    def run(self):
+        '''Initialise the runner function with passed args, kwargs.'''
+        # Retrieve args/kwargs here; and fire processing using them
+        try:
+            self.signals.started.emit()
+            result = self.fn(*self.args, **self.kwargs)
+        except:
+            traceback.print_exc()
+            exctype, value = sys.exc_info()[:2]
+            self.signals.error.emit((exctype, value, traceback.format_exc()))
+        else:
+            # Return the result of the processing
+            self.signals.result.emit(result)
+        finally:
+            self.signals.finished.emit()  # Done
